Move BlueZGattService to dbus module
