+++ bless/backends/bluezdbus/server.py
@@ -16,8 +16,8 @@ from bleak.backends.bluezdbus.characteristic import (  # type: ignore
 
 from bless.backends.server import BaseBlessServer  # type: ignore
 from bless.backends.bluezdbus.application import BlueZGattApplication  # type: ignore
-from bless.backends.bluezdbus.service import BlueZGattService  # type: ignore
 from bless.backends.bluezdbus.characteristic import flags_to_dbus
+from bless.backends.bluezdbus.dbus.service import BlueZGattService  # type: ignore
 from bless.backends.bluezdbus.dbus.utils import get_adapter  # type: ignore
 from bless.backends.bluezdbus.dbus.characteristic import (  # type: ignore
     BlueZGattCharacteristic,
